--- build.gradle
@@ -1,34 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'io.github.sarhatabaot'
-version '0.1.0-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-    maven { url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/' }
-    maven { url = 'https://oss.sonatype.org/content/repositories/snapshots' }
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.3.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.3.1'
-    compile group: 'org.json', name: 'json', version: '20090211'
-    compile group: 'org.jetbrains', name: 'annotations', version: '16.0.1'
-    compileOnly 'org.spigotmc:spigot-api:1.13.2-R0.1-SNAPSHOT'
-}
-
-configurations {
-    shade
-    compile.extendsFrom shade
-}
-jar {
-    configurations.shade.each { dep ->
-        from(project.zipTree(dep)) {
-            exclude 'META-INF', 'META-INF/**'
-        }
-    }
-}
