+++ src/main/java/io/github/sarhatabaot/farminassistreboot/Config.java
@@ -1,4 +1,4 @@
-package io.github.sarhatabaot;
+package io.github.sarhatabaot.farminassistreboot;
 
 import java.io.File;
 import java.io.IOException;
